generalizes a facets into maybe_facet_wrap
tries to fix subplot title offsets ... this is weird, at first I thought you
just have to adjust for the column the plot is in, but the amount to adjust
depends on the number of rows, too.

It seems like the more rows there are in a plot, the less drastic you have to push
the subplot title on the xaxis (to the right) as you increase the column of the
subplot.